Fix for windows JIRA SB-1061
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@60573 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objtools/blast/seqdb_reader/test/seqdb_perf.cpp
@@ -175,7 +175,7 @@ CSeqDBPerfApp::x_ScanDatabase()
     } // end of omp parallel
 
     sw.Stop();
-    uint64_t bases = static_cast<uint64_t>(num_letters / sw.Elapsed());
+    Uint8 bases = static_cast<Uint8>(num_letters / sw.Elapsed());
     cout << "Scanning rate: " 
          << NStr::NumericToString(bases, NStr::fWithCommas) 
          << " bases/second" << endl;
